[1.2.0] Split `Config` class
